refactor: Separated relations from schema
