+++ project/lib/db/schema.ts
@@ -1,7 +1,6 @@
 // TODO: Task 3.1 - Design database schema for users, projects, lists, and tasks
 // TODO: Task 3.3 - Set up Drizzle ORM with type-safe schema definitions
 
-import { relations } from "drizzle-orm";
 import {
   boolean,
   index,
@@ -168,76 +167,5 @@ export const projectTeams = pgTable(
     index("project_teams_creator_idx").on(table.isCreator),
   ]
 );
-
-// Define relations
-export const usersRelations = relations(users, ({ many }) => ({
-  teamMembers: many(teamMembers),
-  ledTeams: many(teams, { relationName: "team_leader" }),
-  assignedTasks: many(tasks, { relationName: "task_assignee" }),
-  createdTasks: many(tasks, { relationName: "task_creator" }),
-  createdProjects: many(projects, { relationName: "project_creator" }),
-  comments: many(comments, { relationName: "comment_author" }),
-}));
-
-export const teamsRelations = relations(teams, ({ many, one }) => ({
-  members: many(teamMembers),
-  projects: many(projectTeams),
-  leader: one(users, {
-    fields: [teams.leaderId],
-    references: [users.id],
-    relationName: "team_leader",
-  }),
-}));
-
-export const tasksRelations = relations(tasks, ({ many, one }) => ({
-  project: one(projects, {
-    fields: [tasks.projectId],
-    references: [projects.id],
-  }),
-  assignee: one(users, {
-    fields: [tasks.assigneeId],
-    references: [users.id],
-    relationName: "task_assignee",
-  }),
-  createdBy: one(users, {
-    fields: [tasks.createdById],
-    references: [users.id],
-    relationName: "task_creator",
-  }),
-  comments: many(comments),
-}));
-
-export const projectsRelations = relations(projects, ({ many, one }) => ({
-  teams: many(projectTeams),
-  tasks: many(tasks),
-  createdBy: one(users, {
-    fields: [projects.createdById],
-    references: [users.id],
-    relationName: "project_creator",
-  }),
-}));
-
-export const projectTeamsRelations = relations(projectTeams, ({ one }) => ({
-  project: one(projects, {
-    fields: [projectTeams.projectId],
-    references: [projects.id],
-  }),
-  team: one(teams, {
-    fields: [projectTeams.teamId],
-    references: [teams.id],
-  }),
-}));
-
-export const commentsRelations = relations(comments, ({ one }) => ({
-  task: one(tasks, {
-    fields: [comments.taskId],
-    references: [tasks.id],
-  }),
-  author: one(users, {
-    fields: [comments.authorId],
-    references: [users.id],
-    relationName: "comment_author",
-  }),
-}));
 
 // Placeholder exports to prevent import errors
